refactor(web): filter-panel の共有 props 型を types.ts へ切り出す
子から index への type import で生じる循環参照を避け、依存方向を単方向に保つ。
