+++ application/packages/web/src/client/features/article/components/filter-panel/index.tsx
@@ -1,13 +1,7 @@
 import { useIsMobile } from '@/client/components/shadcn/hooks/use-mobile'
-import { type FilterParams } from '../../hooks/use-articles'
 import { DesktopFilterPanel } from './desktop-filter-panel'
 import { MobileFilterPanel } from './mobile-filter-panel'
-
-export interface FilterPanelProps {
-  applied: FilterParams
-  onApplyFilters: (filters: FilterParams) => void
-  isLoggedIn: boolean
-}
+import { type FilterPanelProps } from './types'
 
 export function FilterPanel({ applied, onApplyFilters, isLoggedIn }: FilterPanelProps) {
   const isMobile = useIsMobile()
